Changed seed type to unsigned int.

+++ src/dataset/crossvalidation_adaptator.hpp
@@ -155,7 +155,7 @@ namespace dataset {
 
     class CrossValidationProperties {
     public:
-        CrossValidationProperties(const DataFrame& df, int k, long unsigned int seed, bool include_null) : k(k),
+        CrossValidationProperties(const DataFrame& df, int k, unsigned int seed, bool include_null) : k(k),
                                                                                              m_seed(seed),
                                                                                              indices(), 
                                                                                              limits(), 
@@ -205,7 +205,7 @@ namespace dataset {
         friend class CrossValidation;
     private:
         int k;
-        long unsigned int m_seed;
+        unsigned int m_seed;
         std::vector<int> indices;
         std::vector<const_vecit<int>> limits;
         bool include_null;
@@ -215,7 +215,7 @@ namespace dataset {
     public:
         CrossValidation(const DataFrame df, int k, bool include_null = false) : CrossValidation(df, k, std::random_device{}(), include_null) { }
 
-        CrossValidation(const DataFrame df, int k, long unsigned int seed, bool include_null = false) : 
+        CrossValidation(const DataFrame df, int k, unsigned int seed, bool include_null = false) : 
                                                     m_df(df),
                                                     prop(std::make_shared<CrossValidationProperties>(m_df, k, seed, include_null)) { }
