+++ src/factors/continuous/SemiparametricCPD.hpp
@@ -56,7 +56,7 @@ namespace factors::continuous {
         }
 
         Array_ptr sample(int n, const DataFrame& evidence_values, 
-                         long unsigned int seed = std::random_device{}()) const;
+                         unsigned int seed = std::random_device{}()) const;
 
         std::string ToString() const;
