+++ src/pybindings/pybindings_learning/pybindings_scores.cpp
@@ -96,7 +96,7 @@ void pybindings_scores(py::module& root) {
         .def(py::init<const DataFrame&, int>(),
                 py::arg("df"),
                 py::arg("k") = 10)
-        .def(py::init<const DataFrame&, int, long unsigned int>(),
+        .def(py::init<const DataFrame&, int, unsigned int>(),
                 py::arg("df"),
                 py::arg("k") = 10,
                 py::arg("seed"))
@@ -106,7 +106,7 @@ void pybindings_scores(py::module& root) {
         .def(py::init<const DataFrame&, double>(),
                 py::arg("df"),
                 py::arg("test_ratio") = 0.2)
-        .def(py::init<const DataFrame&, double, long unsigned int>(),
+        .def(py::init<const DataFrame&, double, unsigned int>(),
                 py::arg("df"),
                 py::arg("test_ratio") = 0.2,
                 py::arg("seed"))
